Added sp_helplogins which is a system stored procedure that provides information about logins (both SQL and Windows) and users associated with those logins.
sp_helplogins [ [ @LoginNamePattern = ] N'LoginNamePattern' ] [ ; ]
@LoginNamePattern is sys.sysname, with a default of NULL . If specified, @loginnamepattern must exist.

It returns two result sets -
First view - information about all logins present on the server.
Second view - information about each login and its mapping with a user in a database or its membership with a database-role

Task - BABEL-5742

+++ contrib/babelfishpg_tsql/sql/upgrades/babelfishpg_tsql--5.2.0--5.3.0.sql
@@ -219,7 +219,7 @@ BEGIN
     BEGIN    
         SELECT DISTINCT
             CAST(LExt.orig_loginname AS sys.SYSNAME) AS LoginName,
-            CAST(CAST(Base.oid AS INT) AS sys.varbinary(85)) AS SID,
+            CAST(CAST(Base.oid AS BIGINT) AS sys.varbinary(85)) AS SID,
             CAST(LExt.default_database_name AS SYS.SYSNAME) AS DefDBName,
             CAST(LExt.default_language_name AS SYS.SYSNAME) AS DefLangName,
             CASE 
@@ -303,7 +303,7 @@ BEGIN
 
         SELECT DISTINCT
             CAST(LExt.orig_loginname AS sys.SYSNAME) AS LoginName,
-            CAST(CAST(Base.oid AS INT) AS sys.varbinary(85)) AS SID,
+            CAST(CAST(Base.oid AS BIGINT) AS sys.varbinary(85)) AS SID,
             CAST(LExt.default_database_name AS SYS.SYSNAME) AS DefDBName,
             CAST(LExt.default_language_name AS SYS.SYSNAME) AS DefLangName,
             CASE
